move orbital projection to proc; rename orthogonalize_orbitals.py to orbital_helpers.py
